Add Framework
Added framework and separation of responsibilities. May want to
extrapolate the framework into it's own project, but for now we will let
it sit here. The idea is that the framework should provide, well, a
framework for handling things a NodeMCU should be doing... like setting
up MQTT brokers, setting up a webserver for HTTP APIs, connecting to a
hot spot and/or acting as an access point.